Add more eslint rule
