+++ .prettierrc
@@ -1,0 +1,34 @@
+{
+    "tabWidth": 4,
+    "useTabs": false,
+    "singleQuote": true,
+    "semi": true,
+    "trailingComma": "es5",
+    "bracketSpacing": true,
+    "bracketSameLine": false,
+    "arrowParens": "avoid",
+    "embeddedLanguageFormatting": "auto",
+    "printWidth": 140,
+    "overrides": [
+        {
+            "files": ["*.ts"],
+            "options": {
+                "parser": "typescript"
+            }
+        },
+        {
+            "files": ["*.json"],
+            "options": {
+                "useTabs": true,
+                "parser": "json5"
+            }
+        },
+        {
+            "files": ["*.YML"],
+            "options": {
+                "tabWidth": 2,
+                "parser": "yaml"
+            }
+        }
+    ]
+}
