--- .prettierrc.json
@@ -1,26 +1,0 @@
-{
-	"tabWidth": 4,
-	"useTabs": false,
-	"singleQuote": true,
-	"semi": true,
-	"trailingComma": "es5",
-	"bracketSpacing": true,
-	"bracketSameLine": false,
-	"arrowParens": "avoid",
-	"embeddedLanguageFormatting": "auto",
-	"printWidth": 140,
-	"overrides": [
-		{
-			"files": "*.json",
-			"options": {
-				"useTabs": true
-			}
-		},
-		{
-			"files": "*.YML",
-			"options": {
-				"tabWidth": 2
-			}
-		}
-	]
-}
