+++ eslint.config.mjs
@@ -26,70 +26,24 @@ export default tseslint.config(
             '@typescript-eslint/consistent-type-definitions': ['off', 'type'],
             '@typescript-eslint/no-explicit-any': ['error', { ignoreRestArgs: true, fixToUnknown: true }],
             '@typescript-eslint/no-misused-promises': ['error', { checksVoidReturn: false }],
+            'no-duplicate-imports': ['error'],
+            'no-constructor-return': ['error'],
+            'no-inner-declarations': ['error'],
+            'no-self-compare': ['error'],
+            'no-template-curly-in-string': ['error'],
+            'no-unmodified-loop-condition': ['error'],
+            'no-unreachable-loop': ['error'],
+            'arrow-body-style': ['error', 'as-needed'],
+            'block-scoped-var': ['error'],
+            curly: ['error'],
+            'default-case-last': ['error'],
+            'default-param-last': ['error'],
+            eqeqeq: ['error'],
+            'grouped-accessor-pairs': ['error'],
+            'max-classes-per-file': ['error', { max: 1 }],
+            'max-depth': ['error', { max: 2 }],
+            'max-lines': ['error', { max: 500, skipComments: true, skipBlankLines: true }],
+            'no-else-return': ['error', { allowElseIf: true }],
         },
     }
 );
-
-// {
-//     rules: {
-//         'no-console': ['error'],
-//             '@typescript-eslint/no-var-requires': ['off'],
-//             '@typescript-eslint/no-explicit-any': ['off'],
-//             'no-duplicate-imports': ['error'],
-//             'no-constructor-return': ['error'],
-//             'no-inner-declarations': ['error'],
-//             'no-promise-executor-return': ['error'],
-//             'no-self-compare': ['error'],
-//             'no-template-curly-in-string': ['error'],
-//             'no-unmodified-loop-condition': ['error'],
-//             'no-unreachable-loop': ['error'],
-//             'no-useless-assignment': ['error'],
-//             'arrow-body-style': ['error', 'as-needed'],
-//             'block-scoped-var': ['error'],
-//             camelcase: [
-//             'error',
-//             {
-//                 properties: 'always',
-//                 ignoreImports: true,
-//                 ignoreDestructuring: false,
-//             },
-//         ],
-//             complexity: [
-//             'error',
-//             {
-//                 max: 2,
-//             },
-//         ],
-//             curly: ['error'],
-//             'default-case-last': ['error'],
-//             'default-param-last': ['error'],
-//             eqeqeq: ['error'],
-//             'grouped-accessor-pairs': ['error'],
-//             'max-classes-per-file': [
-//             'error',
-//             {
-//                 max: 1,
-//             },
-//         ],
-//             'max-depth': [
-//             'error',
-//             {
-//                 max: 2,
-//             },
-//         ],
-//             'max-lines': [
-//             'error',
-//             {
-//                 max: 500,
-//                 skipComments: true,
-//                 skipBlankLines: true,
-//             },
-//         ],
-//             'no-else-return': [
-//             'error',
-//             {
-//                 allowElseIf: true,
-//             },
-//         ],
-//     }
-// },
