+++ src/module/feature/events/error.event.ts
@@ -1,8 +1,7 @@
-import type { Client } from 'discord.js';
-import { Events } from 'discord.js';
-import { basename } from 'node:path';
-import { Logger } from '../../shared/utils/logger';
-import type { BotEvent } from './types/bot-event.type';
+import {type Client, Events} from 'discord.js';
+import {basename} from 'node:path';
+import {Logger} from '../../shared/utils/logger';
+import type {BotEvent} from './types/bot-event.type';
 
 const logger: Logger = new Logger(basename(__filename));
