+++ src/module/feature/handlers/handlers.handler.ts
@@ -7,12 +7,14 @@ module.exports = async (client: Client): Promise<void> => {
     const handlersDir: string = __dirname;
 
     for (const handler of readdirSync(handlersDir)) {
-        if (!(!handler.endsWith('.ts') || handler.startsWith('handlers'))) {
-            require(`${handlersDir}/${handler}`)(client);
+        if (!handler.endsWith('.ts') || handler.startsWith('handlers')) {
+            continue;
+        }
 
-            if (!EnvUtil.isDev()) {
-                await EnvUtil.sleep(TimeEnum.SECONDE * 10);
-            }
+        require(`${handlersDir}/${handler}`)(client);
+
+        if (!EnvUtil.isDev()) {
+            await EnvUtil.sleep(TimeEnum.SECONDE * 10);
         }
     }
 };
