+++ src/module/feature/slash-commands/models/trivia.model.ts
@@ -10,146 +10,53 @@ import {
     ComponentType,
     EmbedBuilder,
     type Message,
-    StringSelectMenuBuilder,
-    type StringSelectMenuInteraction,
-    StringSelectMenuOptionBuilder,
 } from 'discord.js';
 import type { Ammo } from '../../../shared/apis/wot/models/wot-api.type';
 import { LoggerInjector } from '../../../shared/decorators/injector/logger-injector.decorator';
-import { Singleton } from '../../../shared/decorators/injector/singleton-injector.decorator';
-import { Table } from '../../../shared/decorators/injector/table-injector.decorator';
-import { EmojiEnum } from '../../../shared/enums/emoji.enum';
 import { TimeEnum } from '../../../shared/enums/time.enum';
 import type { TriviaSelected } from '../../../shared/singleton/trivia/models/trivia.type';
-import type { TriviaSingleton } from '../../../shared/singleton/trivia/trivia.singleton';
 import type { TriviaAnswer } from '../../../shared/tables/complexe-table/players-answers/models/players-answers.type';
-import type { PlayersAnswersTable } from '../../../shared/tables/complexe-table/players-answers/players-answers.table';
 import type { TriviaPlayer } from '../../../shared/tables/complexe-table/players/models/players.type';
-import type { PlayersTable } from '../../../shared/tables/complexe-table/players/players.table';
 import type { Tank } from '../../../shared/tables/complexe-table/tanks/models/tanks.type';
 import type { TriviaData } from '../../../shared/tables/complexe-table/trivia-data/models/trivia-data.type';
-import type { TriviaDataTable } from '../../../shared/tables/complexe-table/trivia-data/trivia-data.table';
 import type { WinStreak } from '../../../shared/tables/complexe-table/win-streak/models/win-streak.type';
-import type { WinStreakTable } from '../../../shared/tables/complexe-table/win-streak/win-streak.table';
-import { DateUtil } from '../../../shared/utils/date.util';
-import type { Logger } from '../../../shared/utils/logger';
 import { StringUtil } from '../../../shared/utils/string.util';
 import { TimeUtil } from '../../../shared/utils/time.util';
-import { MEDAL } from '../../../shared/utils/variables.util';
 import { ShellEnum, ShellType } from '../enums/shell.enum';
 import type { PlayerAnswer } from '../types/trivia-game.type';
+import { TriviaStatisticsModel } from './trivia-statistics.model';
 
 @LoggerInjector
-export class TriviaModel {
-    //region INJECTABLE
-    private readonly logger: Logger;
-    @Singleton('Trivia') private readonly trivia: TriviaSingleton;
-    @Table('TriviaData') private readonly triviaTable: TriviaDataTable;
-    @Table('Players') private readonly playersTable: PlayersTable;
-    @Table('PlayersAnswer') private readonly playerAnswerTable: PlayersAnswersTable;
-    @Table('WinStreak') private readonly winStreakTable: WinStreakTable;
-    //endregion
-
-    //region PRIVATE READONLY FIELDS
-    /**
-     * Embed containing all the information for the rule command
-     */
-    private readonly embedRule: EmbedBuilder = new EmbedBuilder()
-        .setColor(Colors.Orange)
-        .setTitle('Voici les règles concernant le jeu trivia V2.1')
-        .setThumbnail(
-            'https://img.poki.com/cdn-cgi/image/quality=78,width=600,height=600,fit=cover,f=auto/8fe1b52b0dce26510d0ebf4cbb484aaf.png'
-        )
-        .setFields(
-            {
-                name: 'But',
-                value: "Ce jeu vise à t'aider à mémoriser les dégâts moyens et le type d'obus des chars de rang 10 dans World of Tanks.",
-            },
-            {
-                name: 'Commande',
-                value: "Le jeu `trivia` peut-être lancé avec la commande `/trivia game` dans n'importe quel salon textuel. Toutefois, il ne peut pas être lancé que `4 fois par jour`. Lorsque tu démarres un trivia, le bot t'envoie un message visible uniquement par toi contenant les informations suivantes :",
-            },
-            {
-                name: 'Obus',
-                value: `Affichant son \`type\` (AP, APCR, etc) et son \`dégât moyen (alpha)\`, l'obus peut être un obus \`standard\` ou un obus \`spécial\` (dit gold). **${EmojiEnum.WARNING} Depuis la V2.1, le bot récupère le canon dit 'méta' des chars. Cependant, il n'utilise qu'un seul canon, donc faites attention au char comme le E-100 ou c'est le premier canon qui est sélectionné !**`,
-            },
-            {
-                name: 'Minuteur',
-                value: `Tu as \`30 secondes\` pour répondre à la question. À la fin du minuteur, le bot t'enverra ton résultat : bonne ou mauvaise réponse ainsi que les informations sur le char à trouver et sur le char que tu as sélectionné.\n\nLorsque tu répond à la question en moins de \`10 secondes\`, tu obtiens un bonus de \`25%\` sur les points obtenus en cas de bonne réponse. ${EmojiEnum.WARNING} **Le temps de réponse change si tu sélectionnes une autre réponse !**`,
-            },
-            {
-                name: 'Bouton',
-                value: `Le message sera suivi de \`quatre boutons cliquables\`. Chaque bouton représente un char rang 10 sélectionné aléatoirement. Pour répondre, il te suffit de cliquer sur l'un des boutons. Tu peux changer de réponse tant que le minuteur n'est pas terminé. **${EmojiEnum.WARNING} ️Quand 2 ou plusieurs chars ont le même obus (type et alpha), tous ces chars sont considérés comme la réponse.**`,
-            },
-            {
-                name: 'Sommaire',
-                value: 'Tous les jours, au lancement du bot, un sommaire sera envoyé. Il contient le top 3 des joueurs pour chaque question en terme de vitesse de réponse, ainsi que la bonne réponse et des informations sur les autres chars.',
-            },
-            {
-                name: 'AFK',
-                value: "En cas d'absence de jeu pendant une journée, une perte de `1.8% de vos points` sera appliquée.",
-            },
-            {
-                name: 'Valeur',
-                value: `${EmojiEnum.WARNING} Les valeurs indiquées dans les paragraphes précédent peuvent être modifiées sans avoir été mis à jour dans le texte. Une communication pourra être fais dans ce cas la !`,
-            }
-        );
-    /**
-     * Embed use as exemple for the rule command
-     */
-    private readonly embedExample: EmbedBuilder = new EmbedBuilder()
-        .setTitle('Example de question')
-        .setDescription("Dans cette exemple, les boutons sont clickable mais aucune logique n'est implémentée !")
-        .setColor(Colors.Blurple)
-        .setFields(
-            {
-                name: 'Obus :',
-                value: `\`${ShellEnum.ARMOR_PIERCING} 390\``,
-                inline: true,
-            },
-            {
-                name: 'Minuteur :',
-                value: 'Le temps sera ici',
-                inline: true,
-            }
-        );
-    /**
-     * The row builder for the example
-     */
-    private readonly rowExample: ActionRowBuilder<ButtonBuilder> = new ActionRowBuilder<ButtonBuilder>()
-        .addComponents(new ButtonBuilder().setCustomId('Object 140').setLabel('Object 140').setStyle(ButtonStyle.Primary))
-        .addComponents(new ButtonBuilder().setCustomId('Manticore').setLabel('Manticore').setStyle(ButtonStyle.Primary))
-        .addComponents(new ButtonBuilder().setCustomId('Object 268').setLabel('Object 268').setStyle(ButtonStyle.Primary))
-        .addComponents(new ButtonBuilder().setCustomId('Object 907').setLabel('Object 907').setStyle(ButtonStyle.Primary));
-    //endregion
-
+export class TriviaModel extends TriviaStatisticsModel {
     //region PRIVATE FIELDS
     /**
      * @see TriviaData.max_duration_of_question
      * @unit minute
      */
     private maxQuestionDuration: TriviaData['max_duration_of_question'];
+
     /**
      * @see TriviaData.max_response_time_limit
      * @unit second
      */
     private responseTimeLimit: TriviaData['max_response_time_limit'];
+
     /**
      * @see TriviaData.max_number_of_question
      */
     private maxNumberOfQuestion: TriviaData['max_number_of_question'];
+
     /**
      * The map of all players with there metadata
      */
-    private datum: Map<
+    private datum = new Map<
         string,
         {
             player: TriviaPlayer;
             questionNumber: number;
             interaction: ChatInputCommandInteraction;
         }
-    >;
-
+    >();
     //endregion
 
     /**
@@ -159,16 +66,6 @@ export class TriviaModel {
         this.maxQuestionDuration = TimeEnum.MINUTE * (await this.triviaTable.getMaxDurationOfQuestion());
         this.responseTimeLimit = TimeEnum.SECONDE * (await this.triviaTable.getMaxResponseTimeLimit());
         this.maxNumberOfQuestion = await this.triviaTable.getMaxNumberOfQuestion();
-        this.datum = new Map();
-    }
-
-    /**
-     * Callback for the rule commande, send the rule to the player in ephemeral message
-     *
-     * @param {ChatInputCommandInteraction} interaction - The chat interaction of the player
-     */
-    public async sendRule(interaction: ChatInputCommandInteraction): Promise<void> {
-        await interaction.editReply({ embeds: [this.embedRule, this.embedExample], components: [this.rowExample] });
     }
 
     /**
@@ -181,21 +78,13 @@ export class TriviaModel {
 
         if (!player) {
             try {
-                const added = await this.playersTable.addPlayer(interaction.user.username);
-
-                if (!added) {
-                    throw new Error('Return false');
-                }
-
-                this.logger.info('New player added to trivia database {}', interaction.user.username);
-                player = await this.playersTable.getPlayerByName(interaction.user.username);
+                player = await this.addPlayer(interaction.user.username);
             } catch (reason) {
                 await interaction.editReply({
                     content:
                         'Il y a un problème avec la base de données, merci de réessayer plus tard. Si le problème persist merci de contacter <@313006042340524033>',
                 });
                 this.logger.error(`Failed to insert new player in database with error :`, reason);
-                return;
             }
         }
 
@@ -270,170 +159,6 @@ export class TriviaModel {
         this.logger.debug('Trivia game message send to {}', interaction.user.username);
 
         this.collectAnswer(gameMessage, interaction.user.username);
-    }
-
-    /**
-     * Sends the trivia statistics for a specific month to the user.
-     *
-     * @param {ChatInputCommandInteraction} interaction - The interaction object representing the user command.
-     */
-    public async sendStatistics(interaction: ChatInputCommandInteraction): Promise<void> {
-        const player: TriviaPlayer = await this.playersTable.getPlayerByName(interaction.user.username);
-
-        if (!player) {
-            await interaction.editReply({
-                content: "Tu n'as pas encore joué à Trivia. Essaye après avoir répondu au moins une fois au jeu. (`/trivia game`)",
-            });
-            return;
-        }
-
-        const lastAnswer: TriviaAnswer = await this.playerAnswerTable.getLastAnswerOfPlayer(player.id);
-
-        if (!lastAnswer) {
-            await interaction.editReply({
-                content:
-                    "Tu n'as pas encore de statistiques pour le jeu Trivia. Essaye après avoir répondu au moins une fois au jeu. (`/trivia game`)",
-            });
-            return;
-        }
-
-        const periods: Date[] = (await this.playerAnswerTable.getAllPeriodsOfPlayer(player.id))
-            .map(({ month, year }) => {
-                const date = new Date();
-
-                date.setMonth(month - 1);
-                date.setFullYear(year);
-
-                return date;
-            })
-            .reverse();
-
-        const select: StringSelectMenuBuilder = new StringSelectMenuBuilder()
-            .setCustomId('trivia-statistics-select')
-            .setPlaceholder('Choisissez un mois');
-
-        periods.forEach((date: Date) =>
-            select.addOptions(
-                new StringSelectMenuOptionBuilder()
-                    .setLabel(`${date.toLocaleDateString('fr-FR', { month: 'long', year: 'numeric' })}`)
-                    .setValue(`${date.toISOString()}`)
-            )
-        );
-
-        const row = new ActionRowBuilder<StringSelectMenuBuilder>().addComponents(select);
-        const message: Message<BooleanCache<CacheType>> = await interaction.editReply({
-            components: [row],
-            content: 'Choisissez un mois pour voir les statistiques.',
-        });
-
-        message
-            .createMessageComponentCollector({ componentType: ComponentType.StringSelect, time: TimeEnum.HOUR })
-
-            .on('collect', async (i: StringSelectMenuInteraction): Promise<void> => {
-                const date = new Date(i.values[0]);
-
-                const stats: TriviaAnswer[] = await this.playerAnswerTable.getPeriodAnswerOfPlayer(player.id, date);
-                const winStreak: WinStreak = await this.winStreakTable.getWinStreakFromDate(player.id, date);
-
-                const embed = new EmbedBuilder()
-                    .setTitle(
-                        `Statistiques pour le mois de ${date.toLocaleDateString('fr-FR', {
-                            month: 'long',
-                            year: 'numeric',
-                        })}`
-                    )
-                    .setColor(Colors.LuminousVividPink)
-                    .setDescription('Voici les statistiques demandées')
-                    .setFields(
-                        { name: 'Elo', value: `\`${stats[stats.length - 1].elo}\``, inline: true },
-                        {
-                            name: 'Nombre de bonnes réponses',
-                            value: `\`${stats.reduce((number: number, { right_answer }) => {
-                                return number + (right_answer ? 1 : 0);
-                            }, 0)}\``,
-                            inline: true,
-                        },
-                        { name: 'Plus longue séquence correcte', value: `\`${winStreak.max}\`` },
-                        {
-                            name: 'Réponse la plus rapide',
-                            value: `\`${Math.min(...stats.filter(({ answer_time }): boolean => answer_time !== null).flatMap(({ answer_time }): number => answer_time!)) / TimeEnum.SECONDE}\` sec`,
-                            inline: true,
-                        },
-                        {
-                            name: 'Réponse la plus longue',
-                            value: `\`${Math.max(...stats.filter(({ answer_time }): boolean => answer_time !== null).flatMap(({ answer_time }): number => answer_time!)) / TimeEnum.SECONDE}\` sec`,
-                            inline: true,
-                        },
-                        {
-                            name: 'Nombre de participation',
-                            value: `\`${stats.filter(({ trivia_id }) => trivia_id !== null).length}\``,
-                        }
-                    );
-
-                await i.update({
-                    embeds: [embed],
-                });
-            });
-    }
-
-    /**
-     * Sends the scoreboard for the current month's trivia game.
-     *
-     * @param {ChatInputCommandInteraction} interaction - The interaction object representing the user's command interaction.
-     *
-     * Update here with database
-     */
-    public async sendScoreboard(interaction: ChatInputCommandInteraction): Promise<void> {
-        const username = interaction.user.username;
-        const embedScoreboard = new EmbedBuilder()
-            .setTitle('Scoreboard')
-            .setDescription(`Voici le scoreboard du mois de \`${DateUtil.getCurrentMonth()}\` pour le jeu trivia`)
-            .setFooter({ text: 'Trivia game' })
-            .setColor(Colors.Fuchsia);
-
-        const players: TriviaPlayer[] = await this.playersTable.getAllPlayers();
-
-        if (players.length === 0) {
-            await interaction.editReply({
-                content: "Aucun joueur n'a pour l'instant joué au jeu",
-            });
-            return;
-        }
-
-        const promises: Promise<(TriviaAnswer & TriviaPlayer) | null>[] = players.map(({ id }) =>
-            this.playerAnswerTable.getLastAnswerWithPlayerOfPlayer(id)
-        );
-
-        let playerStats = (await Promise.all(promises)).filter(
-            (value: (TriviaAnswer & TriviaPlayer) | null): boolean => value !== null
-        ) as Awaited<TriviaAnswer & TriviaPlayer>[];
-
-        const today = new Date();
-        playerStats = playerStats.filter(({ date }) => date.getMonth() === today.getMonth() && date.getFullYear() === today.getFullYear());
-
-        if (playerStats.length === 0) {
-            await interaction.editReply({
-                content: "Aucun joueur n'a pour l'instant joué ce mois si au jeu",
-            });
-            return;
-        }
-
-        playerStats.sort((a: TriviaAnswer & TriviaPlayer, b: TriviaAnswer & TriviaPlayer) => b.elo - a.elo);
-
-        embedScoreboard.addFields({
-            name: 'Joueur - Elo',
-            value: playerStats
-                .map(
-                    (player: TriviaAnswer & TriviaPlayer, index: number): string =>
-                        `${username === player.name ? '`--> ' : ''}${index < 3 ? MEDAL[index] : index + 1}. ${player.name} - ${player.elo}${username === player.name ? ' <--`' : ''}`
-                )
-                .join('\n'),
-            inline: true,
-        });
-
-        await interaction.editReply({
-            embeds: [embedScoreboard],
-        });
     }
 
     /**
@@ -544,11 +269,8 @@ export class TriviaModel {
         const { allTanks, selectedTanks } = this.getData(username);
 
         const isGoodAnswer = this.isGoodAnswer(playerAnswer, username);
-
         const answerEmbed: EmbedBuilder = this.createAnswerEmbed(true, selectedTanks.tank, isGoodAnswer);
-
         const otherAnswer: EmbedBuilder[] = [];
-
         const ammo: Ammo = selectedTanks.tank.ammo[selectedTanks.ammoIndex];
 
         allTanks.forEach((vehicle: Tank): void => {
@@ -678,9 +400,7 @@ export class TriviaModel {
         }
 
         const { allTanks, selectedTanks } = this.getData(playerName);
-
         const lastAnswer: TriviaAnswer = await this.playerAnswerTable.getLastAnswerOfPlayer(value.player.id);
-
         const oldElo = lastAnswer?.elo ?? 0;
         const elo = this.calculateElo(oldElo, playerAnswer?.responseTime, isGoodAnswer);
 
@@ -711,18 +431,7 @@ export class TriviaModel {
 
         if (!winStreak) {
             try {
-                const added = await this.winStreakTable.addWinStreak(value.player.id);
-
-                if (!added) {
-                    await playerAnswer.interaction.editReply({
-                        content: 'Une erreur est survenue lors de la création de tes statistiques dans la base de données !',
-                    });
-                    this.logger.error(StringUtil.transformToCode(`Failed to create winstreak for player {} in database`, playerName));
-                    return;
-                }
-
-                this.logger.info('Successfully create winstreak for player {} in database', playerName);
-                winStreak = { max: 0, current: 0 };
+                winStreak = await this.addWinStreak(playerAnswer.interaction, playerName, value.player.id);
             } catch (reason) {
                 await playerAnswer.interaction.editReply({
                     content: 'Une erreur est survenue lors de la création de tes statistiques dans la base de données !',
@@ -798,10 +507,7 @@ export class TriviaModel {
                     .setDescription('Bravo a trouvée la bonne réponse')
                     .setColor(Colors.Green)
                     .setFooter({ text: 'Trivia Game' })
-                    .setFields({
-                        name: 'Elo',
-                        value: `Ton nouvelle elo est : \`${newElo}\` (modification de \`${newElo - oldElo}\`)`,
-                    }),
+                    .setFields({ name: 'Elo', value: `Ton nouvelle elo est : \`${newElo}\` (modification de \`${newElo - oldElo}\`)` }),
             ],
         });
         this.logger.debug('Player {} found the right answer', playerName);
@@ -859,11 +565,7 @@ export class TriviaModel {
                     .setFooter({ text: 'Trivia Game' })
                     .setImage(tank.image)
                     .setFields(
-                        {
-                            name: 'Char sélectionné',
-                            value: `\`${tank.name}\``,
-                            inline: true,
-                        },
+                        { name: 'Char sélectionné', value: `\`${tank.name}\``, inline: true },
                         {
                             name: "Catégorie d'obus",
                             value: `\`${datum.ammoIndex ? ShellType.GOLD : ShellType.NORMAL} \``,
@@ -874,10 +576,7 @@ export class TriviaModel {
                             value: `\`${ShellEnum[ammo.type as keyof typeof ShellEnum]} ${ammo.damage[1]}\``,
                             inline: true,
                         },
-                        {
-                            name: 'Elo',
-                            value: `Ton nouvelle elo est : \`${newElo}\` (modification de \`${newElo - oldElo}\`)`,
-                        }
+                        { name: 'Elo', value: `Ton nouvelle elo est : \`${newElo}\` (modification de \`${newElo - oldElo}\`)` }
                     ),
             ],
         });
@@ -909,5 +608,48 @@ export class TriviaModel {
             );
             return;
         }
+    }
+
+    /**
+     * Add the player to the database
+     *
+     * @param {string} playerName - The player name
+     *
+     * @return {TriviaPlayer} - The player if the put is successful
+     *
+     * @throws {Error} - If the sql query failed
+     */
+    private async addPlayer(playerName: string): Promise<TriviaPlayer> {
+        const added = await this.playersTable.addPlayer(playerName);
+
+        if (!added) {
+            throw new Error('Return false');
+        }
+
+        this.logger.info('New player added to trivia database {}', playerName);
+        return await this.playersTable.getPlayerByName(playerName);
+    }
+
+    /**
+     * Create the win streak line in the database for the given player
+     *
+     * @param {ButtonInteraction} interaction - The interaction to replay if error
+     * @param {string} playerName - The player name
+     * @param {number} playerId - The id of the player in the database
+     *
+     * @return {WinStreak} - The win streak object initialise
+     */
+    private async addWinStreak(interaction: ButtonInteraction, playerName: string, playerId: number): Promise<WinStreak> {
+        const added = await this.winStreakTable.addWinStreak(playerId);
+
+        if (!added) {
+            await interaction.editReply({
+                content: 'Une erreur est survenue lors de la création de tes statistiques dans la base de données !',
+            });
+            throw new Error(StringUtil.transformToCode(`Failed to create winstreak for player {} in database`, playerName));
+        }
+
+        this.logger.info('Successfully create winstreak for player {} in database', playerName);
+        return { max: 0, current: 0 } as WinStreak;
     }
 }
