+++ src/module/shared/utils/random.util.ts
@@ -14,22 +14,22 @@ export class RandomUtil {
     /**
      * Create an array of number fill with random number between {@link max} and {@link min}
      *
+     * @param {boolean} [allowRepeat] - Whether repetition of numbers is allowed in the array.
+     * @param {number[]} [forbidden] - An array of numbers that are forbidden to be generated.
      * @param {number} [length=1] - The length of the array to generate.
      * @param {number} [max=1] - The maximum value (inclusive) of the random numbers.
      * @param {number} [min=0] - The minimum value (inclusive) of the random numbers.
-     * @param {boolean} [allowRepeat] - Whether repetition of numbers is allowed in the array.
-     * @param {number[]} [forbidden] - An array of numbers that are forbidden to be generated.
      *
      * @throws {Error} - Will throw an error if the length is less than 1 or if max is less than min.
      *
      * @return {number[]} - An array of random integers within the specified range.
      */
     public static getArrayWithRandomNumber(
+        allowRepeat?: boolean,
+        forbidden?: number[],
         length: number = 1,
         max: number = 1,
-        min: number = 0,
-        allowRepeat?: boolean,
-        forbidden?: number[]
+        min: number = 0
     ): number[] {
         if (!Number.isInteger(length) || length < 1) {
             throw new Error('Length must be a positive integer greater than 0.');
